pyproject.toml
[tool.poetry]
name = "askupload"
version = "0.1.0"
description = ""
authors = ["ThomasB91 <116346781+ThomasB91@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
st-weaviate-connection = {git = "https://github.com/weaviate/st-weaviate-connection.git"}
pip = "^24.0"
install = "^1.3.5"
python-dotenv = "^1.0.1"
openai = "^1.20.0"
instructor = "^1.2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
